Add extra feature related to qaoa for weighted maxcut
Specifically, the function InitializeVectorAsWeightedMaxCutCostFunction()
has been added. It receives a BaseType vector as adjacency matrix instead
of the 1/0 integer vector of InitializeVectorAsMaxCutCostFunction().
Furthermore, it returns the max weight of cut edges.
The weighted case does not work well with GetHistogramFromCostFunction()
which assumes an integer max_value and transform all cut values to integers
in {0, 1, 2, ..., max_cut}.