Fix bug in PermuteGlobalQubits() and in corresponding unit tests
For the unit test, notice that when EXPECT_* is used in a single rank
(possibly different from the main rank 0), even if the condition is not
satisfied the test is recorded as successful. This may be due to the fact
that in 'suite_of_tests.cpp' we listen to the main rank only.

Using ASSERT_* in ranks different than 0 still collapse the unit test
and is therefore more robust to MPI problems.