Update MPI init and finalize in Python bindings
For naming consistency, the MPI environment is initialized via:
  iqs.EnvInit()
and the finalization via:
  iqs.EnvFinalize()

when the number of ranks is not a power of 2 or, in general, when some
of the ranks are dummy, we need to finalize them and terminate their execution
before proceding with the actual program.
The Python examples '/notebooks/python_mpi.py' and '/notebooks/python_script.py'
include appropriate code.