Add example using PermuteQubits()
In the example 'communication_reduction_via_qubit_reordering.cpp',
we simulate a simple circuit (with gates on the last 10 qubits only)
using the trivial qubit ordering and the inverse one.
With increasing number of MPI processes, the relative cost of the
simulation with inverse qubit order (w.r.t. the simulation with
trivial qubit order) decreases.
This is expected since no MPI communication is needed for the inverse order.